Fix uninstall may failed in KensingtonWorks
Prior to this change, I have used to run the uninstaller that comes with
KensingtonWorks to uninstall the related files. But I discovered the
uninstaller always require an user to press the "Uninstall" button to
continue the process, which may cause the terminal to wait for the
script completed.

Unfortunately, there is no ways to uninstall the cask silently. Since
there is no `-h` or `--help` switches to show any arguments it accepts.
Then I tried to extract the strings from the uninstaller binary and
found some shell scripts to remove the files.