xppen-pentablet 3.2.3_211203. Version rollback and fix.
The previous version of this cask pointed to a non-existent future version. The URL pointed to the correct (latest) version. As a result, Homebrew was able to download the .dmg file, but it couldn't install it.

This change also fixes the installation of drivers. The suite's  directory contains a hidden subdirectory with supporting files (drivers). If the .app is copied in isolation, it is unable to install drivers upon first start.

Confirmed working by testing in a fresh MacOS installation.