Add canon-mf-ufr2-printer v10.19.11
Canon-mf-printer cask does not cover all Canon MF printers.  This cask
is meant to provide drivers for printers that are not included in the
canon-mf-printer cask.

The printers covered by this driver include but are not limited by:
MF74x, MF64x, MF84xx, MF9xxx, MF1xxx series printers.